pyproject.toml
[tool.poetry]
name = "vub"
version = "0.1.0"
description = "Implementation of the experiments in the paper 'Explainable Regularization in DNNs Using the Information Bottleneck' (2023)"
authors = ["Nir Weingarten <nirzvi89@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.7.9"
sklearn = "<=1.0.2"
scikit-learn = "1.0.2"
torchvision = "0.14.1"
transformers = "4.29.2"
wandb = "^0.15.8"
dill = "^0.3.7"
pickle5 = "0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
